requirement.txt
arrow==1.3.0
asgiref==3.8.1
bcrypt==4.1.2
binaryornot==0.4.4
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==42.0.5
cycler==0.12.1
Django==5.0.4
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-rest-framework==0.1.0
django-tailwind==3.8.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
fonttools==4.51.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
pillow==10.3.0
psycopg==3.1.18
pycparser==2.22
Pygments==2.17.2
PyJWT==1.7.1
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.0
sshtunnel==0.4.0
text-unidecode==1.3
threadpoolctl==3.4.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
zipp==3.18.1
